pyproject.toml
[project]
name = "scrapling-job-board-scrapper"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "langchain-core>=1.3.3",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "postgrest>=2.30.0",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "scrapling[fetchers]>=0.4.7",
    "supabase>=2.30.0",
]
